+++ lib/minitest_to_rspec/exp/calls/twice.rb
@@ -1,0 +1,11 @@
+require_relative "../call"
+require_relative "../../errors"
+
+module MinitestToRspec
+  module Exp
+    module Calls
+      class Twice < Call
+      end
+    end
+  end
+end